Move the data.c and data.h files to lang.c and lang.h
This is in preparation for another change for toybox: as of right now,
static constant data is interwoven all together with functions when I
cut a toybox release. That won't do, so I want to put all of the
constant data into one file. I wanted to call it data.c, but that was
already taken.

However, I realized that the current data.c really holds language
contructs (for the bc language), so that is where the rename comes from.