Remove the geline I/O stuff
It turns out that toybox requires POSIX 2008, which has getline, so we
an use it. So I am removing the code that I put in (copied from
somewhere).

The function bc_io_fread() was written by me, though, and it is still
needed. It was renamed bc_fread() and it was moved to bc.c.