Fix a bug in handling the math library
The bug was that when parsing the library, bc was going until
BC_STATUS_LEX_EOF, which is an actual error now. Because an error
happened, the parse and program data were "reset," which in this case
means that the execution context was set to the first instruction after
the assignment to scale in the library, which made "a(1)" produce "0"
even though scale was supposed to be 20.