Make printing code only something that debug versions have
In order to use it now, a programmer will have to insert it, like
inserting printf statements. I thought that doing it that way was better
because people who just want the software do not have to worry about an
extra command-line option, and it simplifies code just a touch.

Also, it allows me to eliminate it from toybox, making Landley happier.