Attempt to fix crashes caused by the AFL fuzzer
After inspecting the files used to crash it, I am pretty sure that at
least one source of crashing is receiving '\0' bytes from input (because
of inputting binary files), which is causing fgets() to not do what I
wanted (though it is doing the right thing). I replaced fgets() with a
loop and fgetc() instead.

This change also has another advantage: checking for binary files is not
in the lexer, but in bc_io_getline(), which will automatically stop the
loop in bc_stdin() without explicitly checking for the status
BC_STATUS_LEX_BIN_FILE.