Fix wild memory allocations in sine
This was a wild bug. As it turns out, it was one thing wrong that
manifested itself in a completely different area.

The bug was that the parser was putting the wrong index for a function
when the function did not exist yet. That was the bug.

What this meant was that when the user ran sine, which has a call to
arctangent before arctangent was defined, the call to arctangent would
get replaced with a call to sine itself. This created an infinite
recursion. Also, on a call to sine, it multiplies the scale. This meant
that the scale was increasing on each recursion, causing it to increase
without bound, which in turn, would cause more and more memory to be
requested since scale determines the size of numbers.

Like I said, this was a bug that manifested itself in strange ways.

And sine *still* doesn't work. (It just doesn't eat memory.)