Fix infinite loop problem in sine
The problem was in multiplication. All procedures expect that a number
that is zero should have zero length and 0 radix. It turns out that when
a multiplication returned zero because the result was too small for the
scale, it was not being cleared correctly; it was only being cleared to
the radix.

To fix the problem, I simply had it clear until a non-zero, or the
length was zero. If the length was zero, then the number was zero, so
the radix would be cleared as well. If not, and the length was less than
the radix, the length would be set to the radix.

Oh, and the sign would be cleared as well, if necessary.