Don't have randmath.sh automatically add tests to the suite
The reason for this is that many tests are cases where this bc is off by
only one digit, and that is because of the numerical instability of the
sine algorithm. I have decided to accept an occasional wrong digit in
the last place.

What this means is that I want to manually review all of the test cases
that might be added. That is what this change allows. It goes through
all of the test cases that this bc failed and lets me manually review
them after running the test loop.