Remove memory leaks on errors
These memory leaks were tough. As it turns out, it was because of a
design bug. When the lexer encounters a name, number, or string, it
allocates memory for the data. The parser or VM must then free it. That
was a bad thing.

The solution was to allocate a buffer once upfront, and then, when the
parser or VM needed the data, it would allocate enough space and copy
the data. Then that data would be owned by the entity that allocated the
space.

That made the memory management easier. I ran the entire test suite
(including the errors) through valgrind, and it all passed.