Add a test found by afl
This test is a bit different. It actually found a ***HUGE*** bug that
was obvious in retrospect. The bug was that when doing binary operators
with two variables or two array elements, if the second var/array hadn't
existed before the operation, the var or array vectors might be
expanded, which includes a realloc(). Since the first var/element was
already found, its num pointer would became invalid.

Of course, this caused segfaults, invalid reads, double frees, the whole
shebang (thank you, Valgrind).

This commit only adds the test that afl used to find it, which even
requires changes to errors.sh. Next commit will have the fix.