Fix the bug that I described in the last commit message
This is not the cheapest fix; it basically just refinds the var/array.
However, despite its cost, I like it for several reasons:

1) It will always work. This is because since the var/element have
   already been found, it will exist and not expand the vectors, which
   would invalidate the pointer to the second var/element.
2) It is simple. This commit adds one branch and one function call.
3) It uses code that I already have a lot of confidence in.
4) It would not happen that often anyway. To get it to happen, you have
   to add a lot of variables/elements and manage to trigger the expand
   *after* the first variable/element is already gotten. Most people
   don't even use bc, and when they do, the scripts are usually not
   long.

The cost really isn't that high anyway. It consists of a binary search,
an allocation, and a free. The binary search wouldn't really be a
problem anyway, and in general, the allocation will be pretty fast,
especially since the first search required that same memory too. That
should mean that allocators might, in general, be able to reuse the
memory that was just used, and then free it again.