Change the sqrt solution
The other was a hack. This one is less so. What it does is it detects
when a loop has repeated 4 times. When that happens, it increases the
scale of intermediate operations by 1. That way, evetually (and still
somewhat quickly), it should resolve the stalemate and come to a
solution.