Remove memset() calls where possible
I ran perf on bc when it was running the test suite, and surprisingly,
memset() was the most used procedure! It turns out that I still had a
lot of superfluous memset() calls from when I was writing the code.

While it didn't reduce runtime as much as I had hoped, it did make a
small dent.