Change all of bc to output error messages at point of error
This necessitated several changes, most of which are based on an idea
from busybox.

The first change is a split between statuses and errors. This made a lot
of things cleaner, but basically touched every source and header file.
There were changes to bc_vm_error() and bc_vm_posixError() to handle
this change, and then a few scattered changes to make it all work.