Reduce memory usage
I have a lot to explain with this commit. It is going in on a request
from Denys Vlasenko, who ran "yes 1 | bc" and noticed that there was no
deduplication of constant numbers and that memory grew linearly, without
bound. (This was besides the basically negligible memory growth from the
growing generated code.) It is an issue that I have known about for a
while, but I didn't think it was terribly necessary to solve it.

However, on his request for an idea, I decided to try an idea that I
came up with right away: using the current map infrastructure (used for
checking for duplicate function names, which means that functions are
redefined) on constants. Adding it for strings too was easy enough, so I
did that too.

The result is that strings and constants are deduplicated, though at a
cost of extra runtime. The extra runtime comes from the binary search
that is performed when inserting (or attempting to insert) into a map.

Surprisingly, it made the parse test unusable. I never did get it to
finish. I mean, it was slow before, but it didn't even finish anymore.
That makes sense; the parse test was millions of lines long, and each
of those lines was a unique number, so the deduplication was hitting
its worst case every time. After thinking about it, I decided that its
size is probably completely unnecessary; the test should be thorough
enough with less numbers (looping to 4096 instead of 65536).

So that was why I reduced the length of the generated parse test in the
last commit.

The end result is that this bc is slightly (exacerbated by the millions
of numbers in the parse test and by the fast, but not constant-time, of
O(log(n)) for binary search) slower in parsing, but as of right now, the
biggest hog of CPU time is still the math. I thought that it was a
worthy tradeoff. My bc can still handle copious amounts of numbers
without failing; it is just slower at doing so, but it also uses less
memory for repeated numbers (and slightly more if no numbers are
repeated). And a lot of numbers are bound to be repeated, like "0" and
"1", and in fact, many already are in the internal math library, so it
should be a net win.

This change may not stay, but as of right now, I am happy with it.