Parse statement delimiters better
The busybox maintainer had a comment in his code that mentioned the
inconsistency of how they were parsed, as well as a test that his bc
failed. He was right on the inconsistency, though I don't know if my bc
would have failed the test because I never tested that before adding
this change. It passes that test now, though, and I will add an
equivalent test in the next commit.