Fix a bug in ubytes()
This wasn't a bug, per se. What was happening is that l(x) (divided by
l(2), of course) was returning a number slightly over 64 for 2^64-1,
because of precision issues. In this commit, scale is just set to a
ridiculously high number (100; because we are dealing with integers,
that is plenty), making the function accurate enough.