Add fixed-width integer functions to lib2
I also made it so they printed error messages if certain conditions were
not met. This makes it a little bit more explicit for the user. Also, I
sped up ubytes() by an order of magnitude by using a loop instead of
directly calling l2(), which calls the expensive l().