Fix a crash/logic bug
The problem was that I was simply copying a number into ibase/obase,
instead of making sure that it was truncated as the spec requires. This
was also causing a crash because bc sometimes would have to realloc the
space for BcProgram::ib or BcProgram::ob, which are static arrays and
cannot be realloc'ed. Just converting the unsigned long val to the
proper BcNum is the correct fix for both problems.