Add abs() as a built in
This commit also has a refactoring in bc_program_builtin() aimed to
remove the bc_program_builtins[] array and reduce executable size.

What was surprising about this commit was that (without the refactoring)
it only added 8 bytes to the executable size, and it actually removed
some after the refactoring.

***WARNING***: The refactoring code takes advantage of code that is not
committed yet. This commit is broken. I had to commit it this way to
sort out the commits.