Remove the use of readlink in install.sh
After investigating this, I found that readlink is not specified by
POSIX. However, that does not mean that it was the better solution, as
the use of `ls -dl` and `printf ${L#*-> }` might not have been portable.
The possible issue was the use of `-> ` to get at the link target.

That said, after reading the POSIX standard for `ls`, I found out that
the `-> ` syntax is required by the standard if `ls` is called with the
option `-l` and without the option `-L`. That means that my original
formulation is most portable.