Implement scientific and engineering notations
These are only available when extra math is enabled because they are
implemented using shifting.

This bc/dc now accepts scientific/engineering notation for input, using
a lowercase 'e' as the exponent start. Either positive or negative
exponents are allowed, and the notation does not have to be correct
(i.e., both programs are liberal in what they accept, whether the
decimal point is in the correct position or not).

For outputting in scientific notation, set `obase` to `0`. This was used
because `0` is not a valid `obase` in other contexts, and doing it this
way instead of a function allows dc to have access to the same feature.
For outputting in engineering mode, set `obase` to `1` (which is also
not valid in other contexts).

For dc, negative exponents use underscore ('_').