Fix EOF behavior in interactive mode with history
Previously, EOF would re-process the previous line, read one more
line, then exit. This is because bc_history_line left the buffer
untouched in case of EOF, and the bc_vm_stdin loop reads the next
line before checking the `done` condition of the previous iteration.

This fix also prevents reading past the end of the buffer if the
first user input is EOF, since in that case vec->len is still 0,
so (size_t)-1 is passed to bc_read_binary.