Commit an adjusted patch sent to me by Stefan Eßer
This patch starts a branch exploring the possibility of using a more
packed representation of digits in BcNum's. I am not sure this will work
out, but it can't hurt to try, especially since Stefan was so kind to
come up with a patch and send it.

As of right now, the things that are known to be broken:

* Printing decimal numbers
* Power (seg fault)