Replace Newton-Raphson algorithm by Goldschmidt algorithm.
For now, keep the Newton-Raphson algorithm in an #if 0 block for reference and to
allow to compare them in further tests.

I have noticed segmentation faults when dividing very large numbers, which probably
are due to missing calls of bc_num_extend() for destination variables of arithmetic
operations. I'll try to fix the length calculations, but I do not know the exact
semantics of a number of low level calls.