Add correction to result of Goldschmidt	algorithm
The correction is derived from the difference of the input parameter times it inverse.
This value could be used for a final Newton-Raphson step (at the cost of an additional
multiplication), but since we can assume that the input	parameter was very near	to 1.0,
the multiplication would only add non-zero digits beyond the current scale range.

The Goldschmidt	algorithm is an	infinite series	of ever smaller positive values	and
if the series is cut off at some point,	the result is known to be exact	or smaller
than the exact value.

The correction applied is meant	to compensate for the neglected	series elements.
If it is found that there are boundary cases where we over-compensate, then the
Newton-Raphson step could be used to apply a slightly smaller correction.