Optimize division even more
There are two changes.

1) I cut the subtractions per iteration nearly in half by just
subtracting from cpb.
2) I reduced the amount of division by making it so division is only
calculated to the current scale, rather than to the end of cpa.

These changes chop about a third off the test suite. It's still slower
than with char-based digits, but it is closer. I might try to optimize
more, but I am satisfied with this if I can't do any more.