Fix the crash found by afl
This crash was caused by parsing empty if statements wrong. There was a
bit of stuff wrong.

First, when if statements were retired, their labels were not set
sometimes.

Second, if there was a brace after an empty if statement, the flags
vector was not properly popped, so the brace stayed, even though it
should not have.

Third, when parsing a single body, I should not end a body, if it has
already been ended.

All of the above are fixed in this commit.