Implement faster print function for obase != 10
This code is only compiled in if compiled with "-D USE_SE_PRINT".

The conversion of the integer part is significantly faster than with the
standard print function (by more than a factor of 2) for large numbers.
There is no measurable difference for small numbers.

All BcDig values in the variable to be printed have been prepared in such
a way, that printing could be performed from num[0] to num[i-1]. IOW, there
is no need to reverse the whole number in a BcVec, it could be done in
separate chunks for each BcDig.