Add a new feature and some tests for it
This feature is the ability to query for the max ibase, obase, and/or
scale supported by the implementation. This allows scripts to stop after
the max is reached, for example.

The test is really an extra print test that prints beyond obase == 99.
It just uses if statements to print (and prints equivalent strings
otherwise).

I also added them to the manuals and added POSIX errors.