Revert "Add an optimization idea from Stefan"
This reverts commit 10e2989e1c701e154b3393d3e2d1cb6a1a7e29be.

It turns out that this is slightly *slower* on average. I think the
reason is because since the quotient biased high, everything biased high
and thus, made it so less iterations of the innermost loop had to
happen.

Another reason could be one that I saw while debugging: if subtracting
once put the number below the target, it often happened that in the next
iteration, with pow /= 10, it basically added back that amount most of
the time, wasting a lot of work, whereas in the old version, it would
subtract, realize that that was too far, back up, and start again with
subtracting less, thus only doing a few operations.