Fix pathological behavior in case integer is rounded up a lot
If the top-most BcDig contains a small value and there are lower
non-zero BcDigs, the integer divisor is incremented to prevent
division results that do not fit into a BcDig (are > BC_BASE_POW).

E.g. for 1000 / 1.1 the integer divisor will become 2 and the first
estimate of the result will be 1000 / 2 = 500, giving a reminder of
1000 - 500 * 1.1 = 450. The next estimate will be 700 = 500 + 450 / 2
and so on with the reminder slowly decreasing in each iteration.

Since the relative error due to the approximation of the full division
by an integer division becomes significant for small integer divisors,
it makes sense to extend the divisor if it i ssmall. In fact, it seems
to pay off for all divisors that are < BC_BASE_POW / 10.