Implement caching of numeric constants in vectors of BcNums
This implementation is not optimized for speed, yet. I have plans to
add copy-on-write semantics for BcNums (i.e. add a reference count and
copy the BcNum only if its refcount is > 1 it is about to be
modified).

This version fails in the bc/stdin test, but diagnosing this problem
will have to wait for tomorrow ...