Eliminate extra pop instructions
These pop instructions were for removing extra items on the result stack
after assignments. For those cases where we can detect that assignments
are not going to be used, we use special instructions that tell bc not
to put the result of the assignment on the stack.

This took a lot of work, and it did not speed things up as much as I had
hoped. Oh well.