Get rid of new memory leaks
These leaks are because names are not used anymore to identify anything;
the index is computed right in the parser. The leaks were because of
duplicated names not being freed. I knew they existed when I made the
changes.

The best thing about this change is that I was able to get rid of the
extra malloc's by using a buffer in the parser to copy data. That should
free up memory and allocation pressure, making the parser even faster.
(Though it is already fast.)