Do not overallocate BcNum arrays by a factor of 6 to 9
In the worst case (base 36) there are about 5/3 decimal digits per
input character. Since the actual base is not known at that time,
this worst case has to be assumed.

It would be better to delay the actual allocation of the num array to
the point where the constant value is used for the first time, since
the base is known at that point.