Use more than 1 array in references test
Both vidx and nidx were always 0 in bc_program_dereference during the
references test, reducing the effectiveness of the test case.
But even with this change, the use of multi-byte values in the byte
code for vidx and nidx cannot be tested this way (this would require
mode than 256 arrays).

The test still has nidx == 0 and thus does not test variations of that
parameter at all.