Better fix for the long division problem
The problem was that the expansion of a in the calling function was
lost in bc_num_divExtend and not fully restored by expanding by 1.

This caused the first division to operate on a dividend that could
far extend the divisor * BC_BASE_POW leading to a quotient larger
than BC_BASE_POW (which was caught by the assertion check).

The results are identical to those of the previous fix in my tests,
but this fix is more correct in that it makes the operands match
the assumptions made in the algorithm and valid for the passed
arguments due to the preparation of cpa and cpb in the calling
function.

IMHO, bc_num_div_long should be called with plain values of a and
b as passed to bc_num_div and perform the copying and expansions
required itself. That way the preparation of the arguments and the
code that depends on e.g. the relative length of the operands were
next to each other.