Fix and add a test for a bug
This bug is that if a var needed to be pushed, and it wasn't last, the
var was pulled before the type check happened. This caused an assert to
(rightly) fire. The type check was moved before, and it all went away.

As said in the tag line, this commit also adds the file generated by afl
that triggered the bug.