Introduce bc_num_addDigits and bc_num_subDigits
The addDigits function is functionally equivalent with addDigit, but
allows simpler code and more error checking. The carry is passed by
reference as an in-out parameter. Passing of 0 as one of the arguments
to addition/subtraction will lead to optimized code and elimination of
redundant operations due to inlining of the small function.

These functions are also slightly more efficient, since they work with
values verified (by assertions) not to overflow a BcDig (and thus do
not depend on BcBigDig variables).

A number of asserts are provided to detect unexpected overflow from
out-of-range parameters. Fuzzing has been performed to make sure that
these assertions are satisfied for all uses of these functions with
all parameters. Care has been taken to make sure that both functions
operate correctly on signed or unsigned BcDig (in case the type was
changed to unsigned at some time).

The subDigits function allows to simplify subArrays and will be used
in combined implementation of bc_num_a and bc_num_s.