Optimize the usual case in bc_num_irand()
The usual case is that the number passed in is less than or equal to
vm->max. In that case, all we have to do is generate one number with
either bc_rand_bounded() or bc_rand_int(). If the bound is equal to
vm->max or is a power of 2, we call bc_rand_int() and mask bits
appropriately. If the bound is *not* a power of 2, we call
bc_rand_bounded().

This change means that only in the case of the bound being greater than
vm->max do we enter the expensive arbitrary-precision generation code.