Optimize BcNum initialization
This commit does several things:

1. It increases the BC_NUM_DEF_SIZE to 32. This is because the smallest
   allocation size on musl is 32, and on glibc, it's 24. This number
   also makes it so on 64-bit, if scale is 20 (as is default with the
   math library), the integer part can be up to 45 digits. This should
   be plenty for most cases. This is important for the items coming
   later.
2. If a number's capacity is less than or equal to BC_NUM_DEF_SIZE * 2,
   it is put onto a list of unused BcNum's.
3. If a new number only needs BC_NUM_DEF_SIZE and there are BcNum's on
   the list of unused BcNum's, one of them is grabbed and used instead.

All of these items together, especially since the default number size is
plenty (and would be the minimum allocation size anyway), means that the
vast majority of uses will just reuse old BcNum's.

This reduces runtime by about 8% on an execution-heavy script, but it
may, in fact, reduce the runtime on math-heavy scripts as well. I will
need to run benchmarks to be sure.