Change the condition for reusing BcNum's
The condition is that the capacity of the BcNum's needs to be exactly
BC_NUM_DEF_SIZE. The reason for this is because in a math-heavy script,
it turns out that allowing BC_NUM_DEF_SIZE * 2 will make bc start
allocating a lot of memory, which makes it take much more time.

Since 32 bytes (either 72 integer digits or 45 integer digits and 27
fractional digits), which is BC_NUM_DEF_SIZE, is going to be enough for
nearly everything, I just decided to make bc only reuse BcNum's that
have exactly that capacity.

The other good side effect of this is that I don't need a special test
for strings masquerading as BcNum's because their capacity is defined to
be 0.