Improve the performance of the interpreter
After a test, I found that `memmove()` when moving memory forward was
faster than adding an extra `setjmp()` and also better than moving
memory backwards. Thus, I decided to change the interpreter to push the
end result on the result stack *before* the retire, making sure that
everything was clear first. Then, the old results are removed during the
retire, and the new result is moved into place.