Remove a line of code that is not necessary
The reason this is not necessary is that BC_SIG_UNLOCK, which always
follows this line of code, does the same thing. The only question is
whether the signal lock can be unlocked before checking for an
exception.

The answer is yes because the signal handler also checks for an
exception, and if one exists, it does nothing. So we don't care about
that space of time.

If no exception exists and a signal interrupts inbetween that time, it
is still not a problem because it will see that no lock is taken and
do the jump itself.