Fix a bug noticed by Stefan Eßer of FreeBSD
This bug was that the stdout buffer was not getting cleared correctly
when signals happened. This was because late in my dev process for
3.0.0, I made a change to remove BC_STATUS_SIGNAL and put it in a
separate variable, but I did not change a condition used to detect the
signal in order to call `bc_program_reset()` which is what cleared the
buffer. This commit changes that condition.