Fix a bug found by Math2 (on GitHub)
This bug is a misuse of strings, specifically `strchr()`. In the `stdin`
buffer, I was using a direct length and foregoing use of a null
character. However, this was a bug because I was also using `strchr()`
to detect newlines in the buffer. This commit just changes it so the
buffer uses a null character.