Fix a crash found by afl in dc
This one requires explanation. To facilitate cleaning up strings and
constants when they are no longer required, I attached them to functions
(BcFunc). This was so I could keep the constants required for functions
in bc even with the garbage collection (bc_vm_clean()). However, in dc,
there really isn't a concept of functions, so cleaning them has always
been a hack. And it resulted in a crash.

This commit changes it so in dc alone, constants and strings are
attached to the program itself. This makes constants easier to clean.
And it turns out that I really shouldn't clean strings and functions
because they are inextricably linked. They also have a map, which would
be difficult to clean.

This commit has several great side effects. First, some possibly
inefficient code was removed from bc_vm_clean(). Second, it cleaned up
the hacks I had to do to support functions and strings in dc. All in
all, it was a commit that I should have put in a long time ago.