Fix a dc crash found by afl
This one was tough. It turns out that constants do need to be attached
to the "function" (string) they are in to prevent bad things from
happening when cleaning out constants in the main function. However,
unlike bc which can load constant lazily, dc cannot because it may put a
constant on the results stack, start executing another string, and
attempt to access that constant in the new string. This is bad.

Now, dc loads constants eagerly. I may make bc do the same.