Fix the crash found in last commit
This crash happened because when assigning strings in dc, there are two
cases: the result has the index of the string, or the number inside of
the result has the index of the string. I forgot this, and only pulled
the index from the result. This commit changes it to pull from the right
place.