Implement a better API for the library
There are several changes.

First, there is a way to garbage collect all of the BcNum's. This was a
suggestion from Stefan.

Second, I changed the `*_create()` functions to consume their arguments.
This was a suggestion from Stefan because he wants to do things like

```
z = libbc_num_add(libbc_num_mul(a, b), libbc_num_div(c, d));
```

which I can understand. However, this caused a problem: how do I return
errors?

I thought about it for a bit and decided to find a way to encode errors
as unused values in BcNumber. So I encoded the errors as

```
0 - (BcNumber) error
```

which means that I can easily convert between the representations, plus
they only occupy the highest indices, which people are unlikely to
reach.

Third, I had to deal with another problem: how do I handle it when
people write code like above, and `libbc_num_div(c, d)` returns an
error? I decided to just propagate errors like floating point does. If
any of the arguments are invalid numbers, invalid numbers are returned.

Fourth, I changed the `*_create()` functions to remove the `_create`
portion of their names and changed the other functions, the ones that
return a BcError, to have `_err` at the end of their names. This was
done to make the `*_create()` functions more compact, better for use as
above. It also makes sense for me to mark the error-returning functions
so that users know they have to handle the errors.