Remove the neg field from BcNum
Whether or not a number is negative is now stored in the zeroth
(right-most) bit of rdx, since the limit on rdx is determined by scale
and since rdx will always be less than half of scale, so there will be,
be definition, an unused bit.

This means that to get the real radix, I have to shift right by 1, and
then I have to shift left by 1 and bitwise or with the neg bit to set
the rdx. I made macros do to all of that.

The reason I chose to make it the zeroth, least significant bit (which,
if set, means the rdx field is 1) is because I once tried to make it the
most significant bit, and there was a much bigger change. Also, it
caused bc to slow down measurably. This change does *not* cause bc to
slow down measurably; all readings are within noise on both. In fact,
the version without the neg field had less variability in the timings.
It is possible that having a neg field is *slightly* faster, but it was
no more than 50 miliseconds over 3 seconds or more.

As an unexpected benefit, this actually shrinks the executable size when
compiling with -Os.

All of the above, plus the fact that arrays of numbers are highly used
in bc and will be used *a lot* in bcl(3) made me decide to keep this
change because it will save one-sixth of the memory when putting BcNum's
into arrays.