Fix another crash
This one was exactly like the other one, except in ocurred in the
*other* use of bc_program_assignStr(). In both cases, I did not
correctly differentiate between the two cases for strings, whether in
the result or in the number (both are possible).

After finding this crash (afl++ found it), I went through all occurences
of strings to see if there were any others, and I could not find any.