Fix a small non-portable thing in configure.sh
This small non-portable thing was doing something I thought was
portable: setting IFS=$'\n'. Brian Callahan of OpenBSD ran into a
problem though, and after I found and fixed the offending bit, I checked
the standard. It is not in the standard for POSIX sh. Sigh...

I can remove this without problem, by the way. That's because it was in
there for safety only, safety for if I used source file names with
spaces, something I haven't done, nor was I planning to. I was just
coding carefully. But since setting IFS is not portable, I have to stick
with file names with no spaces. That's fine, and if I forget, the build
should loudly break.