Make bc/dc not exit with -e and -f in B/DC_ENV_ARGS
This behavior was requested by my one user on NetBSD, and after checking
with Stefan Esser of FreeBSD, we decided that it wasn't a big enough
deal to not do. In fact, in my mind, it fixes a lot of problems with how
B/DC_ENV_ARGS interacts with scripts.