Fix problems with printing stuff without newlines
The problem was not that stuff was printed without newlines, it was that
history did not recognize it as such. This commit fixes that.

To do that, I had a way for bc_file_* functions to know what kind of
flush they needed to do. I added an enum for it and updated all of the
places. This ensures that there isn't an infinite loop or other such
problem when flushing during history.

This commit does not complete the change. I have to ensure it builds
with history turned off still.