Ensure output is flushed in scripts
I want to change this because I was using a print statement to track
progress in a script, and it wasn't flushing. It only flushed when I
sent a SIGINT.

That is not good behavior, and it's behavior I have only because I had
to implement my own flushing for bc.

Fix this by flushing when stuff is printed.