Use cut(1) instead of head(1) with a GNUism in the dc Easter test
When I first added the test, I tried to use cut(1) because it is in
POSIX. However, I couldn't figure out how it worked. (The POSIX standard
was not well written for it.) That was why I used head(1). This patch
changes the test to use cut(1) instead.

This was contributed by the OpenBSD user and contributor Brian Callahan.
Thank you!