Garbage collect temps if memory allocation fails
This commit changes bc_vm_malloc(), bc_vm_realloc(), and bc_vm_strdup()
to call bc_vm_freeTemps() if they fail and try again. This is to make it
so if allocation fails, bc will free memory that it really isn't using
in an attempt to make the allocation succeed.

However, if the allocation fails again, they throw a fatal error.

Signed-off-by: Gavin Howard <gavin@yzena.com>