Make failing to push to the temp vector a non-event
Previously, like all code that allocated, growing the temps vector could
fail, which would cause a fatal error to occur. That should not happen
for a caching operation.

This commit changes the code to try to push, but if it fails, it just
returns the failure. The code that tried to push checks for failure, and
if it did fail, it just frees the temp.

This had a nice side effect: I could also free the temp vector when a
garbage collect had to happen.

Signed-off-by: Gavin Howard <gavin@yzena.com>