Change ops that need integers to accept integers with scale
This was complicated. In order to make it work, I needed to make sure
the numbers appeared like integers to the actual operations. This means
that the numbers had to be adjusted before and after the operations.
This means that I had to make sure that the changes worked with signals
and setjmp/longjmp exceptions.

Signed-off-by: Gavin Howard <gavin@yzena.com>