Fix a crash found by AFL++ and add its test
This crash was caused by invalid free on a fatal memory allocation
error, which I had decided to ignore when BcNum's needed to be unshifted
because of `bc_num_checkInt()`. Turns out, I can't ignore it.

So instead, I decided to do several things:

* Rename `bc_num_checkInt()` to bc_num_nonInt()`
* Make `bc_num_nonInt()` return true if the argument is not an integer.
* If it is an integer, return false and copy the argument to another
  argument, shifting just the copy.

With these steps, the original is never shifted. This has several
benefits, including:

* I don't need to worry about `setjmp()` for these arguments anymore.
* Thus, the code will be correct by default, rather than brittle.

I should have done it this way to start with.

Signed-off-by: Gavin Howard <gavin@yzena.com>