Remove the executable bit from bc and dc scripts
This also necessitated a change to tests/test.sh, in that those scripts
had to be run with the installed bc or dc. This actually made it less
brittle though, and it also means that it will pick up mine when it is
installed in /usr/local. Generating tests should take much less time
when it does that.

Signed-off-by: Gavin Howard <gavin@yzena.com>