Change the build system to add default settings with env vars
This is a big change. First of all, it removes the -P build option, so
prompt is always available now.

Second, it allows the builder to set whether or not certain settings are
on or off by default. These settings are:

* bc.banner (whether bc displays the copyright header in interactive
  mode).
* {bc,dc}.sigint_reset (whether or not bc/dc resets on SIGINT or exits).
* {bc,dc}.tty_mode (whether or not TTY mode is on when available).
* {bc,dc}.history (whether or not history is enabled when available).
* {bc,dc}.prompt (whether prompt is enabled when available).

Then, users can influence the settings of these with environment
variables: BC_BANNER, {BC,DC}_SIGINT_RESET, {BC,DC}_TTY_MODE,
{BC,DC}_HISTORY, and {BC,DC}_PROMPT.

I still have documentation to add, but this commit pretty well does all
of the necessary plumbing in code, except for adding the defaults to the
help texts.

Signed-off-by: Gavin Howard <gavin@yzena.com>