Use slabs for strings
I almost didn't make this change, but after using the new ministat with
the benchmark.sh changes, I found out that, when running all of the
benchmarks together, the slabs version is always faster, after 20 runs
anyway.

What this commit does is add vectors to vm which contain "slabs". These
slabs are a pointer and a length. The pointer always points to an
allocation of 4096 bytes, except when a slab vector is used to allocate
a string that is greater than 128 bytes (including the nul byte). In
that case, space for the string alone is allocated and pushed at the
last spot. That way, I can reduce fragmentation.

Anyway, each slab is basically a bump pointer when allocating space for
strings. When a slab runs out of space, a new one is allocated and used.

This reduces memory allocator pressure by allocating few big pages (in
fact, exactly a page) instead of many small allocations that all need to
be freed. In preliminary tests with Valgrind (which reports the number
of allocations and bytes), these changes could result in 1/4-3/4 as many
allocations, usually 1/4-1/2. Of course, the number of bytes allocated
was usually more, but not by much.

Signed-off-by: Gavin Howard <gavin@yzena.com>